+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Fixed transitive directory dependencies installation.
 - Fixed file permissions for configuration and authentication files.
 - Fixed an error in `cache:clear` for Python 2.7.
+- Fixed publishing for the first time with a prerelease.
 
 
 ## [0.11.5] - 2018-09-04